+++ tests/unit/utils/aws/test_aws_stack.py
@@ -2,24 +2,19 @@ import pytest
 from botocore.utils import InvalidArnException
 
 from localstack.utils.aws.arns import extract_region_from_arn, lambda_function_name, parse_arn
-from localstack.utils.aws.aws_stack import (
-    ENV_ACCESS_KEY,
-    ENV_SECRET_KEY,
-    inject_region_into_env,
-    inject_test_credentials_into_env,
-)
+from localstack.utils.aws.aws_stack import inject_region_into_env, inject_test_credentials_into_env
 
 
 def test_inject_test_credentials_into_env_already_with_none_adds_both():
     env = {}
     inject_test_credentials_into_env(env)
-    assert env.get(ENV_ACCESS_KEY) == "test"
-    assert env.get(ENV_SECRET_KEY) == "test"
+    assert env.get("AWS_ACCESS_KEY_ID") == "test"
+    assert env.get("AWS_SECRET_ACCESS_KEY") == "test"
 
 
 def test_inject_test_credentials_into_env_already_with_access_key_does_nothing():
     access_key = "an-access-key"
-    expected_env = {ENV_ACCESS_KEY: access_key}
+    expected_env = {"AWS_ACCESS_KEY_ID": access_key}
     env = expected_env.copy()
     inject_test_credentials_into_env(env)
     assert env == expected_env
@@ -27,7 +22,7 @@ def test_inject_test_credentials_into_env_already_with_access_key_does_nothing()
 
 def test_inject_test_credentials_into_env_already_with_secret_key_does_nothing():
     secret_key = "a-secret-key"
-    expected_env = {ENV_SECRET_KEY: secret_key}
+    expected_env = {"AWS_SECRET_ACCESS_KEY": secret_key}
     env = expected_env.copy()
     inject_test_credentials_into_env(env)
     assert env == expected_env
